+++ deltachat-jsonrpc/src/lib.rs
@@ -12,23 +12,18 @@ mod tests {
 
     #[tokio::test(flavor = "multi_thread")]
     async fn basic_json_rpc_functionality() -> anyhow::Result<()> {
-        // println!("{}", "");
         let tmp_dir = TempDir::new().unwrap().path().into();
-        println!("tmp_dir: {:?}", tmp_dir);
-
         let accounts = Accounts::new(tmp_dir).await?;
-        let cmd_api = CommandApi::new(accounts);
+        let api = CommandApi::new(accounts);
 
         let (sender, mut receiver) = unbounded::<String>();
 
         let (client, mut rx) = RpcClient::new();
-        let session = cmd_api;
-        let handle = RpcSession::new(client, session);
+        let session = RpcSession::new(client, api);
         tokio::spawn({
             async move {
                 while let Some(message) = rx.next().await {
                     let message = serde_json::to_string(&message)?;
-                    // Abort serialization on error.
                     sender.send(message).await?;
                 }
                 let res: Result<(), anyhow::Error> = Ok(());
@@ -39,7 +34,7 @@ mod tests {
         {
             let request = r#"{"jsonrpc":"2.0","method":"add_account","params":[],"id":1}"#;
             let response = r#"{"jsonrpc":"2.0","id":1,"result":1}"#;
-            handle.handle_incoming(request).await;
+            session.handle_incoming(request).await;
             let result = receiver.next().await;
             println!("{:?}", result);
             assert_eq!(result, Some(response.to_owned()));
@@ -47,7 +42,7 @@ mod tests {
         {
             let request = r#"{"jsonrpc":"2.0","method":"get_all_account_ids","params":[],"id":2}"#;
             let response = r#"{"jsonrpc":"2.0","id":2,"result":[1]}"#;
-            handle.handle_incoming(request).await;
+            session.handle_incoming(request).await;
             let result = receiver.next().await;
             println!("{:?}", result);
             assert_eq!(result, Some(response.to_owned()));